% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.R
\name{effects.fmlogit}
\alias{effects.fmlogit}
\title{Marginal and Discrete Effects of the Covariates}
\usage{
\method{effects}{fmlogit}(object, effect = c("marginal", "discrete"),
  marg.type = "atmean", se = F, varlist = NULL, at = NULL, R = 1000)
}
\arguments{
\item{object}{A "fmlogit" object.}

\item{effect}{Can be "marginal", for marginal effect; or "discrete", for discrete changes from
the min to the max.}

\item{marg.type}{Type of marginal or discrete effects to be computed. Default to "atmean", the effect at 
the mean of all covariates. Also take "aveacr", the averaged effects across all observations. See details.}

\item{se}{Whether to calculate standard errors for those margins. See details.}

\item{varlist}{A string vector which provides the name of variables to calculate 
the marginal effect. If missing, all variables except the constant will be calculated. 
Use "constant" if wish to compute the marginal effect of constant.}

\item{at}{Under development}

\item{R}{Number of times to sample for the Krinsky-Robb standard error. Default to 1000.}
}
\value{
The function returns an object of class "fmlogit.margins". It contains the following component:

\code{effects} A matrix of calculated effects.

\code{se} A matrix of standard errors corresponding to the effects. Shows up if se=T for the 
input parameter.

\code{ztable} A list of matrices containing effects, standard errors, z-stats and p-values.

\code{R} Number of simulation times for Krinsky-Robb standard error calculation. Null if se=F.

\code{expl} String message explaining the effects calculated.
}
\description{
Extract marginal and discrete effects of independent variable from a fractional multinomial logit model.
}
\details{
Interpreting marginal and discrete effects for limited dependent variables can be tricky,
and this is especially true for multinomial logit models. This function provides a wrapper for
estimating discrete and marginal effects out of fractional logit models. 

The marginal effect represents how a unit change in one specific x may influence the choice variable y. 
The estimate of marginal effect is very straighforward. However, special care is needed when summarizing 
the marginal effect across observations. One approach is the estimate the marginal effect while setting
other explanatory variables at the mean. We call this marginal effect at the mean (MEM), which corresponds
to the option \code{marg.type=atmean}. Another approach is to take the average of marginal effects for each
individual. We call this average marginal effect (AME), which corresponds to the option \code{marg.type=
aveacr}. 

The discrete effect represents how a discrete change in one specific x, influence the choice variable y. 
This is more useful for categorical variables, as calculating the "marginal effect" makes little sense
for them. In this function, we calculate the discrete effect by changing the explanatory variable from 
its minimum to its maximum. For a binary variable, this is just the difference between 0 and 1. Similar 
to the marginal effect case, we also have discrete effect at the mean (DEM) and average dscrete effect (ADE).

Standard error is provided for the effects by using Krinsky-Robb(KR) method. Krinsky-Robb is a simulation-based
method that calculates the empirical value of a function given a known distribution of its variables. Here 
we provide Krinsky-Robb standard error for MEM and DEM, and the user can specify how many times of 
simulation \code{R} should the KR algorithm run. 

The user can also specify a subset of explanatory variables when calculating effects. This is done through
specifying string vectors containing the column names of the explanatory variables to \code{varlist}. As the
KR standard error can be time-consuming, it is advised to calculate only the variables in need.
}
\examples{
#results1 = fmlogit(y,X)
effects(results1,effect="marginal")
effects(results1,effect="discrete",varlist = colnames(object$X)[c(1,3)])
}

