% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{fitted.fmlogit}
\alias{fitted.fmlogit}
\alias{predict.fmlogit}
\alias{residuals.fmlogit}
\title{Extract fitted values, residuals, and residuals}
\usage{
\method{fitted}{fmlogit}(object)

\method{residuals}{fmlogit}(object)

\method{predict}{fmlogit}(object, newdata = NULL, newbeta = NULL)
}
\arguments{
\item{object}{A "fmlogit" object.}

\item{newdata}{A new X matrix to perform model prediction. If Null, default to the original dataset. 
X can be a vector with length k, or a matrix with k columns, where k is the number of explanatory 
variables in the original model.}

\item{newbeta}{A new augmented matrix of coefficients that can be used to predict outcome variables. 
Feeds into object$coefficient, which contains the baseline coefficient. Useful for constructing
confidence intervals via simulation or bootstrapping.}
}
\description{
Extract fitted values, residuals, and residuals
}
\examples{
#results1 = fmlogit(y,X)
fitted(results1)
residuals(results1)
predict(results1)
# predict using the first observation from the original dataset.
predict(results1,X[1,])
}

