% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot.fmlogit.margins}
\alias{plot.fmlogit.margins}
\title{"Willingness to Pay" for fmlogit models}
\usage{
\method{plot}{fmlogit.margins}(object, varlist = NULL, X = NULL, y = NULL,
  against = NULL, against.x = NULL, against.y = NULL, group = NULL,
  group.by = NULL, mfrow = NULL)
}
\arguments{
\item{object}{An "fmlogit.margins" object.}

\item{varlist}{A string vector which provides the name of variables to calculate 
the wtp. If missing, all variables in object will be calculated.}

\item{X}{The covariates matrix. Recommend to use element X from the fmlogit object.}

\item{y}{The covariates matrix. Recommend to use element y from the fmlogit object.}

\item{against}{A character string of an index to plot against. Default to Null, and uses
"natural" order of the dataset.}

\item{mfrow}{A two-element numeric vector indicating plot panel row and column. 
Feeds into par(mfrow). Please make sure that the size of the panel is larger than the
number of choices. If NULL, use default method to generate square panels.}
}
\value{
Plots of effects vs. chosen variables
}
\description{
Calculates the willingness to pay for fractional multinomial logit models.
}
\details{

}

