% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_data.R
\docType{data}
\name{spending}
\alias{spending}
\title{Government Spending by Dutch Cities in 2005}
\format{A data frame with 429 row and 12 columns.}
\source{
\href{http://fmwww.bc.edu/repec/bocode/c/citybudget.dta}
}
\usage{
data(spending)
}
\description{
Data from 429 Dutch cities with governmental spending on each sub-category
, and city attributes.
}
\examples{
spending
}
\keyword{datasets}
