% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fmlogit}
\alias{summary.fmlogit}
\alias{summary.fmlogit.margins}
\alias{summary.fmlogit.margins}
\alias{summary.fmlogit.wtp}
\title{Generate summary tables for fmlogit objects}
\usage{
\method{summary}{fmlogit}(object, varlist = NULL, sepline = F, digits = 3,
  add.info = T, list = T, sigcode = c(0.05, 0.01, 0.001), print = F)

\method{summary}{fmlogit.margins}(object, varlist = NULL, sepline = F,
  digits = 3, add.info = T, list = T, sigcode = c(0.05, 0.01, 0.001),
  print = F)

\method{summary}{fmlogit.wtp}(object, varlist = NULL, sepline = F,
  digits = 3, sigcode = c(0.05, 0.01, 0.001), print = F)
}
\arguments{
\item{object}{an object with class "fmlogit", "fmlogit.margins", or "fmlogit.wtp".}

\item{varlist}{select a subset of variable names to be processed. Default to NULL, of which all variables will
be processed.}

\item{sepline}{whether the output table uses separate lines for coefficients and standard errors.}

\item{digits}{number of digits to be signifed. Default to show 3 digits.}

\item{add.info}{whether to add additional descriptive information to the output.}

\item{list}{whether to output a list object, or a single data frame.}

\item{sigcode}{the significance code to be used. Has to be a three-component vector.}
}
\value{
Either a list (for display purposes) or a data.frame (for csv output purposes). If list return (which is
the default) is selected, then the list will contain 4 components: $estimates the estimate; $N number of 
observations, $llf value of the log-likelihood function; and $baseline the name of the baseline choice.
}
\description{
Generate tables of coefficient estimates, partial effects, and willingness to pay from
fmlogit-type objects.
}
\details{
This module provides summary methods for three fmlogit objects: \code{fmlogit}, \code{fmlogit.margins}
, and \code{fmlogit.wtp}. 

The summary method offers several options to the users. The user can choose for a list output \code{list=T}, which is
 good for display and quoting purposes, or a data frame output \code{list=F}, which is good for table outputs. The user
can also specify whether to provide additional information other than the parameter estimates, whether to use 
seperate lines for the estimates and the standard errors (which mimics the output style in Stata),
 as well as the significance code.
}
\examples{
# generate fmlogit summary
#results1 = fmlogit(y,X)

# generate marginal effects summary
#effects1 = effects(results1,effect="marginal")
summary(effects1)

# generate latex style output
# require(xtable)
xtable(summary(effects1,list=F,sepline=T))
}
