% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmlogit.R
\name{fmlogit}
\alias{fmlogit}
\title{Estimate Fractional Multinomial Logit Models}
\usage{
fmlogit(y, X, beta0 = NULL, MLEmethod = "CG", maxit = 50000,
  abstol = 1e-04, ...)
}
\arguments{
\item{y}{the dependent variable (N*J). Can be a matrix or a named data frame. The first column of the matrix is automatically treated as the baseline.}

\item{X}{independent variable (N*K). Can be a matrix or a named data frame. If there is no intercept term in the X, then an intercept term is automatically added.}

\item{beta0}{Initial value for beta used in optimization. Uses a 1*K(J-1) vector. Default to a vector of zeros.}

\item{MLEmethod}{Method of optimization. Goes into \code{optim(method=MLEmethod))}. Default to "CG", the conjugate gradients method.}

\item{maxit}{Maximum number of iteration.}

\item{abstol}{Tolerence.}

\item{...}{additional parameters that goes into \code{optim()}}
}
\value{
\code{estimates} A list of J-1 matrices containing parameter estimates, standard errors, and hypothesis testing results.

\code{likelihood} The likelihood value

\code{convergence} Convergence diagnostics code

\code{count} Dataset information
}
\description{
fmlogit is used to estimate fractional multinomial logit models using quasi-maximum likelihood estimations.
}
\details{
For definition of the fmlogit likelihood function, see Maarten Buis's fmlogit page: http://maartenbuis.nl/software/likelihoodFmlogit.pdf
}
\examples{
fmlogit(y,X)
}

